+++ crates/mako/src/lib.rs
@@ -38,20 +38,8 @@ mod targets;
 #[cfg(test)]
 mod test_helper;
 mod transform;
-mod transform_after_resolve;
-mod transform_async_module;
-mod transform_css_handler;
-mod transform_css_url_replacer;
-mod transform_dep_replacer;
-mod transform_dynamic_import;
-mod transform_env_replacer;
 mod transform_in_generate;
-mod transform_optimizer;
-mod transform_provide;
-mod transform_px2rem;
-mod transform_react;
-mod transform_try_resolve;
-mod transform_virtual_css_modules;
+mod transformers;
 mod tree_shaking;
 mod tree_shaking_analyze;
 mod tree_shaking_module;
