+++ crates/mako/src/plugins/mod.rs
@@ -2,7 +2,7 @@ pub mod assets;
 pub mod copy;
 pub mod css;
 pub mod farm_tree_shake;
-pub(crate) mod hmr_runtime;
+pub mod hmr_runtime;
 pub mod invalid_syntax;
 pub mod javascript;
 pub mod json;
