+++ crates/mako/src/transformers/transform_optimizer.rs
